pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mono-radar"
version = "0.1.0"
description = "Monorepo change-impact analyzer: find what to test/build/deploy from a diff"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "mono-radar contributors"}]
keywords = ["monorepo", "ci", "impact-analysis", "dependency-graph", "build"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
mono-radar = "mono_radar.cli:cli"

[project.urls]
Homepage = "https://github.com/mono-radar/mono-radar"
Issues = "https://github.com/mono-radar/mono-radar/issues"
